+++ MapboxSearch/autofill/src/main/java/com/mapbox/search/autofill/AddressAutofillImpl.kt
@@ -36,7 +36,7 @@ internal class AddressAutofillImpl(
     private val resultFactory: AddressAutofillResultFactory = AddressAutofillResultFactory()
 ) : AddressAutofill, MapboxApiClient {
 
-    override suspend fun reverseGeocoding(
+    override suspend fun reverse(
         point: Point,
         options: AddressAutofillOptions
     ): Expected<Exception, List<AddressAutofillResult>> {