+++ MapboxSearch/place-autocomplete/src/test/java/com/mapbox/search/autocomplete/PlaceAutocompleteImplTest.kt
@@ -164,7 +164,7 @@ internal class PlaceAutocompleteImplTest {
         }
 
         val response = runBlocking {
-            placeAutocomplete.suggestions(TEST_POINT)
+            placeAutocomplete.reverse(TEST_POINT)
         }
 
         assertEquals(TEST_AUTOCOMPLETE_SUGGESTIONS, response.value)
@@ -182,7 +182,7 @@ internal class PlaceAutocompleteImplTest {
         }
 
         val response = runBlocking {
-            placeAutocomplete.suggestions(TEST_POINT)
+            placeAutocomplete.reverse(TEST_POINT)
         }
 
         assertSame(error, response.error)
@@ -213,7 +213,7 @@ internal class PlaceAutocompleteImplTest {
             types = listOf(QueryType.POI, QueryType.ADDRESS, QueryType.STREET),
         )
 
-        runBlocking { placeAutocomplete.suggestions(TEST_POINT, options) }
+        runBlocking { placeAutocomplete.reverse(TEST_POINT, options) }
         coVerify { searchEngine.search(coreOptions) }
         verify(exactly = 1) { activityReporter.reportActivity(eq("place-autocomplete-reverse-geocoding")) }
     }
@@ -229,7 +229,7 @@ internal class PlaceAutocompleteImplTest {
             types = null
         )
 
-        runBlocking { placeAutocomplete.suggestions(TEST_POINT, options) }
+        runBlocking { placeAutocomplete.reverse(TEST_POINT, options) }
 
         assertEquals(ALL_TYPES, slotOptions.captured.types)
     }
@@ -245,7 +245,7 @@ internal class PlaceAutocompleteImplTest {
             types = emptyList()
         )
 
-        runBlocking { placeAutocomplete.suggestions(TEST_POINT, options) }
+        runBlocking { placeAutocomplete.reverse(TEST_POINT, options) }
 
         assertEquals(ALL_TYPES, slotOptions.captured.types)
     }